+++ src/main/java/com/techblog/mapper/ArticleTagMapper.java
@@ -4,5 +4,5 @@ import org.apache.ibatis.annotations.Insert;
 
 public interface ArticleTagMapper {
     @Insert("INSERT INTO tb_article_tags (article_id, tag_id) VALUES (#{articleId}, #{tagId})")
-    void insertArticleTag(Integer articleId, Integer tagId);
+    void insertArticleTag(Long articleId, Long tagId);
 }
