+++ src/main/java/com/techblog/service/IArticleTagService.java
@@ -1,5 +1,5 @@
 package com.techblog.service;
 
 public interface IArticleTagService {
-    void saveArticleTag(Integer articleId, Integer tagId);
+    void saveArticleTag(Long articleId, Long tagId);
 }
