+++ src/main/java/com/techblog/service/impl/UserServiceImpl.java
@@ -163,7 +163,7 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
     @Override
     public Result sign() {
         // 1.获取当前登录用户
-        Integer userId = UserHolder.getUser().getId();
+        Long userId = UserHolder.getUser().getId();
         // 2.获取日期
         LocalDateTime now = LocalDateTime.now();
         // 3.拼接key
@@ -179,7 +179,7 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
     @Override
     public Result signCount() {
         // 1.获取当前登录用户
-        Integer userId = UserHolder.getUser().getId();
+        Long userId = UserHolder.getUser().getId();
         // 2.获取日期
         LocalDateTime now = LocalDateTime.now();
         // 3.拼接key
